+++ src/server/HistoryHandler.java
@@ -6,7 +6,7 @@ import manager.TaskManager;
 import java.io.IOException;
 import java.util.regex.Pattern;
 
-public class HistoryHandler extends BaseHttpHandler {
+public class HistoryHandler extends GenericTypeHandler {
     private final TaskManager manager;
 
     public HistoryHandler(TaskManager manager) {
@@ -14,23 +14,7 @@ public class HistoryHandler extends BaseHttpHandler {
     }
 
     @Override
-    public void handle(HttpExchange httpExchange) throws IOException {
-        try {
-            String requestMethod = httpExchange.getRequestMethod();
-            if (requestMethod.equals("GET")) {
-                getByRequest(httpExchange);
-            } else {
-                System.out.println("Необрабатываемый метод запроса");
-                sendMethodNotAllowed(httpExchange);
-            }
-        } catch (Exception exception) {
-            httpExchange.sendResponseHeaders(CodeResponse.SERVER_ERROR.getCode(), 0);
-        } finally {
-            httpExchange.close();
-        }
-    }
-
-    private void getByRequest(HttpExchange httpExchange) throws IOException {
+    protected void getByRequest(HttpExchange httpExchange) throws IOException {
         String path = httpExchange.getRequestURI().getPath();
         if (Pattern.matches("^/history$", path)) {
             String response = gson.toJson(manager.getHistory());
